Move `Blob` type and methods to separate file
